ABP-7688 refactor
