--- app/uk/gov/hmrc/agentservicesaccount/controllers/ContactDetailsBeforeYouStartController.scala
@@ -1,46 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.agentservicesaccount.controllers
-
-import play.api.i18n.I18nSupport
-import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
-import uk.gov.hmrc.agentservicesaccount.actions.Actions
-import uk.gov.hmrc.agentservicesaccount.config.AppConfig
-import uk.gov.hmrc.agentservicesaccount.connectors.AgentClientAuthorisationConnector
-import uk.gov.hmrc.agentservicesaccount.views.html.pages.contact_details.before_you_start_page
-import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendController
-
-import javax.inject.Inject
-import scala.concurrent.{ExecutionContext, Future}
-
-class ContactDetailsBeforeYouStartController @Inject()(implicit appConfig: AppConfig,
-                                                       agentClientAuthorisationConnector: AgentClientAuthorisationConnector,
-                                                       actions: Actions,
-                                                       beforeYouStartPage: before_you_start_page,
-                                                       ec: ExecutionContext,
-                                                       cc: MessagesControllerComponents) extends FrontendController(cc) with I18nSupport {
-  def showBeforeYouStartPage: Action[AnyContent] = actions.authActionCheckSuspend.async { implicit request =>
-    actions.ifFeatureEnabled(appConfig.enableChangeContactDetails) {
-      if (request.agentInfo.isAdmin) {
-        agentClientAuthorisationConnector.getAgencyDetails().map(agencyDetails =>
-          Ok(beforeYouStartPage(agencyDetails)))
-      } else {
-        Future.successful(Forbidden)
-      }
-    }
-  }
-}
