+++ app/uk/gov/hmrc/agentservicesaccount/controllers/ContactDetailsController.scala
@@ -44,6 +44,7 @@ class ContactDetailsController @Inject()(actions: Actions,
                                          alfConnector: AddressLookupConnector,
                                          evConnector: EmailVerificationConnector,
                                          pcodRepository: PendingChangeOfDetailsRepository,
+                                         agentClientAuthorisationConnector: AgentClientAuthorisationConnector,
                                          //views
                                          view_contact_details: view_contact_details,
                                          check_updated_details: check_updated_details,
@@ -51,7 +52,8 @@ class ContactDetailsController @Inject()(actions: Actions,
                                          update_phone: update_phone,
                                          update_email: update_email,
                                          change_submitted: change_submitted,
-                                         email_locked: email_locked
+                                         email_locked: email_locked,
+                                         beforeYouStartPage: before_you_start_page
                                         )(implicit appConfig: AppConfig,
                                           cc: MessagesControllerComponents,
                                           ec: ExecutionContext) extends FrontendController(cc) with I18nSupport with Logging {
@@ -260,7 +262,7 @@ class ContactDetailsController @Inject()(actions: Actions,
       val arn = request.agentInfo.arn
       sessionCache.get[AgencyDetails](DRAFT_NEW_CONTACT_DETAILS).flatMap {
         case None => // graceful redirect in case of expired session data etc.
-        Future.successful(Redirect(routes.ContactDetailsController.showCurrentContactDetails))
+          Future.successful(Redirect(routes.ContactDetailsController.showCurrentContactDetails))
         case Some(newContactDetails) => for {
           oldContactDetails <- getCurrentAgencyDetails()
           pendingChange = PendingChangeOfDetails(
@@ -282,6 +284,17 @@ class ContactDetailsController @Inject()(actions: Actions,
   val showChangeSubmitted: Action[AnyContent] = actions.authActionCheckSuspend.async { implicit request =>
     ifFeatureEnabled {
       Future.successful(Ok(change_submitted())) // don't show if there is nothing submitted
+    }
+  }
+
+  def showBeforeYouStartPage: Action[AnyContent] = actions.authActionCheckSuspend.async { implicit request =>
+    actions.ifFeatureEnabled(appConfig.enableChangeContactDetails) {
+      if (request.agentInfo.isAdmin) {
+        agentClientAuthorisationConnector.getAgencyDetails().map(agencyDetails =>
+          Ok(beforeYouStartPage(agencyDetails)))
+      } else {
+        Future.successful(Forbidden)
+      }
     }
   }
