+++ app/uk/gov/hmrc/agentservicesaccount/views/pages/contact_details/view_contact_details.scala.html
@@ -58,7 +58,7 @@
     @contact_details_summary_list_partial(agencyDetails, showChangeAction = false)
 
     @if(isAdmin){
-    @p(html = Some(a("view-contact-details.update-details.link", href = routes.ContactDetailsBeforeYouStartController.showBeforeYouStartPage.url)))
+    @p(html = Some(a("view-contact-details.update-details.link", href = routes.ContactDetailsController.showBeforeYouStartPage.url)))
     }
     @p(html = Some(a("view-contact-details.return-to-account.link", href = routes.AgentServicesController.manageAccount.url)))
 }
